autogpts/bosaeed_agent/pyproject.toml
[tool.poetry]
name = "AutoGPT-Forge"
version = "0.1.0"
description = ""
authors = ["Craig Swift <craigswift13@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
tenacity = "^8.2.2"
sqlalchemy = "^2.0.19"
aiohttp = "^3.8.5"
colorlog = "^6.7.0"
chromadb = "^0.4.10"
hypercorn = "^0.14.4"
python-multipart = "^0.0.6"
toml = "^0.10.2"
jinja2 = "^3.1.2"
agbenchmark = { path = "../../benchmark",  develop = true}
weaviate-client = "^3.24.2"
ipykernel = "^6.25.2"
newspaper3k = "^0.2.8"
pandas = "^2.1.1"
bs4 = "^0.0.1"
asyncio = "^3.4.3"
playwright = "^1.38.0"
uvicorn = "^0.23.2"
litellm = "^0.1.821"
duckduckgo-search = "^3.9.3"
selenium = "^4.13.0"
webdriver-manager = "^4.0.1"
watchfiles = "^0.20.0"
[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.3.0"
pre-commit = "^3.3.3"
mypy = "^1.4.1"
flake8 = "^6.0.0"
types-requests = "^2.31.0.2"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
watchdog = "^3.0.0"
mock = "^5.1.0"
autoflake = "^2.2.0"
pydevd-pycharm = "^233.6745.319"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
packages = ["autogpt"]
extend-exclude = '(/dist|/.venv|/venv|/build|/agent|agbenchmark/challenges)/'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip_glob = [".tox", "__pycache__", "*.pyc", "venv*/*", "reports", "venv", "env", "node_modules", ".env", ".venv", "dist", "agent/*", "agbenchmark/challenges/*"]
